+++ PlayCover/Utils/GenshinUserData/RestoreGenshinUserData.swift
@@ -10,37 +10,20 @@ import Foundation
 func restoreUserData(folderName: String, app: PlayApp) {
     let bundleID = app.info.bundleIdentifier
     let isGlobalVersion = bundleID == "com.miHoYo.GenshinImpact"
-    let accountInfoPlistEncrypt = "MIHOYO_ACCOUNT_INFO_PLIST_2_Encryption"
-    let kibanaReportArrayKeyEncrypt = "MIHOYO_KIBANA_REPORT_ARRAY_KEY_Encryption"
-    let lastAccountModelEncrypt = "MIHOYO_LAST_ACCOUNT_MODEL_Encryption"
 
-    let gameDataPath = FileManager.default.homeDirectoryForCurrentUser
-        .appendingPathComponent("Library")
-        .appendingPathComponent("Containers")
-        .appendingPathComponent(bundleID)
-        .appendingPathComponent("Data")
-        .appendingPathComponent("Documents")
-    
-    let storePath = FileManager.default.homeDirectoryForCurrentUser
-        .appendingPathComponent("Library")
-        .appendingPathComponent("Containers")
-        .appendingPathComponent("io.playcover.PlayCover")
-        .appendingPathComponent("Storage")
-        .appendingPathComponent(folderName)
+    let gameDataPath = GenshinUserDataURLs.getGameDataPath(bundleID: bundleID)
+    let storePath = GenshinUserDataURLs.getStorePath(folderName: folderName)
 
-    let accountInfoPlistEncryptUrl = gameDataPath
-        .appendingPathComponent(accountInfoPlistEncrypt)
-    let kibanaReportArrayKeyEncryptUrl = gameDataPath
-        .appendingPathComponent(kibanaReportArrayKeyEncrypt)
-    let lastAccountModelEncryptUrl = gameDataPath
-        .appendingPathComponent(lastAccountModelEncrypt)
+    let accountInfoPlistURL = gameDataPath.appendingPathComponent(GenshinUserDataURLs.accountInfoPlist)
+    let kibanaReportArrayKeyURL = gameDataPath.appendingPathComponent(GenshinUserDataURLs.kibanaReportArrayKey)
+    let lastAccountModelURL = gameDataPath.appendingPathComponent(GenshinUserDataURLs.lastAccountModel)
 
     // Remove existent user data from genshin impact
 
     do {
-        try FileManager.default.removeItem(at: accountInfoPlistEncryptUrl )
-        try FileManager.default.removeItem(at: kibanaReportArrayKeyEncryptUrl)
-        try FileManager.default.removeItem(at: lastAccountModelEncryptUrl)
+        try FileManager.default.removeItem(at: accountInfoPlistURL )
+        try FileManager.default.removeItem(at: kibanaReportArrayKeyURL)
+        try FileManager.default.removeItem(at: lastAccountModelURL)
     } catch {
         Log.shared.log("Error removing file: \(error)")
     }
@@ -48,16 +31,16 @@ func restoreUserData(folderName: String, app: PlayApp) {
     // Move data from StorePath to  GameDataPath
     do {
         try FileManager.default.copyItem(at: storePath
-                                             .appendingPathComponent(accountInfoPlistEncrypt),
-                                         to: accountInfoPlistEncryptUrl)
+            .appendingPathComponent(GenshinUserDataURLs.accountInfoPlist),
+                                         to: accountInfoPlistURL)
 
         try FileManager.default.copyItem(at: storePath
-                                             .appendingPathComponent(kibanaReportArrayKeyEncrypt),
-                                         to: kibanaReportArrayKeyEncryptUrl)
+            .appendingPathComponent(GenshinUserDataURLs.kibanaReportArrayKey),
+                                         to: kibanaReportArrayKeyURL)
 
         try FileManager.default.copyItem(at: storePath
-                                             .appendingPathComponent(lastAccountModelEncrypt),
-                                         to: lastAccountModelEncryptUrl)
+            .appendingPathComponent(GenshinUserDataURLs.lastAccountModel),
+                                         to: lastAccountModelURL)
         if isGlobalVersion {
             let region = try String(contentsOf: storePath
                                                 .appendingPathComponent("region")
@@ -72,19 +55,8 @@ func restoreUserData(folderName: String, app: PlayApp) {
 
 func modifyPlist(newRegion: String) {
     do {
-        // Path of plist file
-        let url = FileManager.default.homeDirectoryForCurrentUser
-            .appendingPathComponent("Library")
-            .appendingPathComponent("Containers")
-            .appendingPathComponent("com.miHoYo.GenshinImpact")
-            .appendingPathComponent("Data")
-            .appendingPathComponent("Library")
-            .appendingPathComponent("Preferences")
-            .appendingPathComponent("com.miHoYo.GenshinImpact")
-            .appendingPathExtension("plist")
-
         // Read plist file
-        let data = try Data(contentsOf: url)
+        let data = try Data(contentsOf: GenshinUserDataURLs.plistPath)
 
         guard var plist = try PropertyListSerialization.propertyList(from: data, options: [], format: nil)
                 as? [String: Any] else { throw PlayCoverError.noGenshinAccount }
@@ -108,13 +80,13 @@ func modifyPlist(newRegion: String) {
         let plistData = try PropertyListSerialization.data(fromPropertyList: plist,
                                                            format: .xml,
                                                            options: 0)
-        try plistData.write(to: url, options: .atomic)
+        try plistData.write(to: GenshinUserDataURLs.plistPath, options: .atomic)
     } catch {
         Log.shared.log("Error editing plist file: \(error)")
     }
 }
 
-func getAccountList () -> [String] {
+func getAccountList() -> [String] {
     let storePath = FileManager.default.homeDirectoryForCurrentUser
         .appendingPathComponent("Library")
         .appendingPathComponent("Containers")
